reconfigured queue files, message is put on queue on file upload success
